# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/invoice_tax_lines:
    get:
      tags:
        - InvoiceTaxLines
      summary: List InvoiceTaxLines
      description: Get a paginated list of InvoiceTaxLines with optional filtering and sorting
      operationId: listInvoiceTaxLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceTaxLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - InvoiceTaxLines
      summary: Create InvoiceTaxLine
      description: Create a new InvoiceTaxLine
      operationId: createInvoiceTaxLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceTaxLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceTaxLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/payment_schedules:
    get:
      tags:
        - PaymentSchedules
      summary: List PaymentSchedules
      description: Get a paginated list of PaymentSchedules with optional filtering and sorting
      operationId: listPaymentSchedules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentScheduleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PaymentSchedules
      summary: Create PaymentSchedule
      description: Create a new PaymentSchedule
      operationId: createPaymentSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentScheduleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    InvoiceTaxLine:
      type: object
      required:
        - id
        - invoice_ref
        - invoice_kind
        - company_id
        - account_id
        - basis
        - taxable_base
        - rate
        - tax_amount
        - exigibility
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        company_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        basis:
          type: string
        description:
          nullable: true
          type: string
          maxLength: 150
        taxable_base:
          type: string
        rate:
          type: string
        tax_amount:
          type: string
        tax_template_id:
          nullable: true
          type: string
          format: uuid
        repartition_line_id:
          nullable: true
          type: string
          format: uuid
        real_account_id:
          nullable: true
          type: string
          format: uuid
        exigibility:
          type: string
        metadata:
          type: object
...
    CreateInvoiceTaxLineRequest:
      type: object
      required:
        - invoice_ref
        - invoice_kind
        - company_id
        - account_id
        - basis
        - taxable_base
        - rate
        - tax_amount
        - exigibility
        - metadata
      properties:
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        company_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        basis:
          type: string
        description:
          nullable: true
          type: string
          maxLength: 150
        taxable_base:
          type: string
        rate:
          type: string
        tax_amount:
          type: string
        tax_template_id:
          nullable: true
          type: string
          format: uuid
        repartition_line_id:
          nullable: true
          type: string
          format: uuid
        real_account_id:
          nullable: true
          type: string
          format: uuid
        exigibility:
          type: string
        metadata:
          type: object
...
    InvoiceTaxLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceTaxLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaymentSchedule:
      type: object
      required:
        - id
        - invoice_ref
        - invoice_kind
        - company_id
        - installment_no
        - due_date
        - amount
        - paid_amount
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        company_id:
          type: string
          format: uuid
        installment_no:
          type: integer
          format: int32
        due_date:
          type: string
          format: date
        amount:
          type: string
        paid_amount:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreatePaymentScheduleRequest:
      type: object
      required:
        - invoice_ref
        - invoice_kind
        - company_id
        - installment_no
        - due_date
        - amount
        - paid_amount
        - status
        - metadata
      properties:
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        company_id:
          type: string
          format: uuid
        installment_no:
          type: integer
          format: int32
        due_date:
          type: string
          format: date
        amount:
          type: string
        paid_amount:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    PaymentScheduleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentSchedule'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []