# src/presentation/http/payment_schedule_handler.rs
//! PaymentSchedule REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use chrono::NaiveDate;
use rust_decimal::Decimal;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::application::service::{PaymentScheduleService, ServiceError};
use crate::domain::entity::*;

// DTO imports
use crate::presentation::dto::{
    CreatePaymentScheduleDto, PatchPaymentScheduleDto, PaymentScheduleResponseDto,
    UpdatePaymentScheduleDto,
};

/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PaymentScheduleError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PaymentScheduleError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PaymentScheduleError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "PAYMENTSCHEDULE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "PAYMENTSCHEDULE_VALIDATION_ERROR"),
            Self::Database(_) => (
                StatusCode::INTERNAL_SERVER_ERROR,
                "PAYMENTSCHEDULE_DATABASE_ERROR",
            ),
            Self::Internal(_) => (
                StatusCode::INTERNAL_SERVER_ERROR,
                "PAYMENTSCHEDULE_INTERNAL_ERROR",
            ),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PaymentSchedule.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /payment_schedules | List with pagination |
/// | POST | /payment_schedules | Create new |
/// | GET | /payment_schedules/:id | Get by ID |
/// | PUT | /payment_schedules/:id | Full update |
/// | PATCH | /payment_schedules/:id | Partial update |
/// | DELETE | /payment_schedules/:id | Soft delete |
/// | POST | /payment_schedules/bulk | Bulk create |
/// | POST | /payment_schedules/upsert | Upsert |
/// | GET | /payment_schedules/trash | List deleted |
/// | POST | /payment_schedules/:id/restore | Restore |
/// | DELETE | /payment_schedules/empty | Empty trash |
/// | GET | /payment_schedules/:id/deleted | Get deleted by ID |
/// | DELETE | /payment_schedules/trash/:id | Permanent delete from trash |
/// | GET | /payment_schedules/count | Count active entities |
/// | GET | /payment_schedules/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PaymentScheduleService::with_repository(repository));
/// let router = create_payment_schedule_routes(service);
/// ```
pub fn create_payment_schedule_routes(service: Arc<PaymentScheduleService>) -> Router {
    BackboneCrudHandler::<
        PaymentScheduleService,
        PaymentSchedule,
        CreatePaymentScheduleDto,
        UpdatePaymentScheduleDto,
        PaymentScheduleResponseDto,
    >::routes(service, "/payment_schedules")
}

/// Create Axum router with only the read (GET) endpoints for PaymentSchedule.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_payment_schedule_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_payment_schedule_read_routes(service: Arc<PaymentScheduleService>) -> Router {
    BackboneCrudHandler::<
        PaymentScheduleService,
        PaymentSchedule,
        CreatePaymentScheduleDto,
        UpdatePaymentScheduleDto,
        PaymentScheduleResponseDto,
    >::read_routes(service, "/payment_schedules")
}

/// Create Axum router with only the write (mutation) endpoints for PaymentSchedule.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_payment_schedule_write_routes(service: Arc<PaymentScheduleService>) -> Router {
    BackboneCrudHandler::<
        PaymentScheduleService,
        PaymentSchedule,
        CreatePaymentScheduleDto,
        UpdatePaymentScheduleDto,
        PaymentScheduleResponseDto,
    >::write_routes(service, "/payment_schedules")
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_payment_schedule_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PaymentScheduleService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_payment_schedule_routes(service).layer(middleware::from_fn(
        move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req
                    .headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| {
                        raw.strip_prefix("Bearer ")
                            .or_else(|| raw.strip_prefix("bearer "))
                    })
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => (
                        axum::http::StatusCode::UNAUTHORIZED,
                        axum::Json(serde_json::json!({
                            "success": false,
                            "error": "unauthorized",
                            "message": "Authentication required"
                        })),
                    )
                        .into_response(),
                }
            }
        },
    ))
}